Extract from methods to referer concern
